qemu_domain.c: do not launch ppc64 guests with APIC-EOI setting
The "<apic/>" feature, although it's not available for pseries,
can be declared in the domain XML of ppc64 guests without errors.
But setting its 'eoi' attribute will break QEMU. For "<apic eoi='on'/>":

qemu-kvm: Expected key=value format, found +kvm_pv_eoi

A similar error happens with eoi='off'.

One can argue that it's better to simply forbid launching ppc64
guests with "<apic/>" declared in the ppc64 XML - it is a feature that
the architecture doesn't support and this would make it clearer
about it. This is sensible, but there are ppc64 guests that are running
with "<apic/>" declared in the domain (and A LOT of guests running with
"<acpi/>" for that matter, probably reminiscent of x86 templates that
were reused for ppc64) that will stop working if we go this route.

A more subtle approach is to detect if the 'eoi' element is being set
for ppc64 guests, and warn the user about it with a better error
message than the one QEMU provides. This is the new error message
when any value is set for the 'eoi' element:

error: unsupported configuration: The 'eoi' attribute of the 'apic'
feature is not supported for architecture 'ppc64' or machine type
'pseries'.

Signed-off-by: Daniel Henrique Barboza <danielhb413@gmail.com>
Message-Id: <20200319214407.714006-2-danielhb413@gmail.com>