qemu_domain.c: do not launch ppc64 guests with 'pmu' setting
The Perfomance Monitoring Unit (PMU) feature is not available for
the Power architecture. The "<pmu/>" feature will always have a value
'on' or 'off' after saving the domain XML, and both will be rejected
by QEMU when launching. This is the error message for
"<pmu state='on'/>":

qemu-kvm: can't apply global host-powerpc64-cpu.pmu=on: Property '.pmu' not found

A similar error message is thrown for "<pmu state='off'/>".

This patch prevents the pseries guest from launching with any
pmu setting with a more informative error message:

error: unsupported configuration: The 'pmu' feature is not
supported for architecture 'ppc64' or machine type 'pseries'

Signed-off-by: Daniel Henrique Barboza <danielhb413@gmail.com>
Message-Id: <20200319214407.714006-4-danielhb413@gmail.com>