conf: allow different resource registration modes
Currently the QEMU driver has three ways of setting up cgroups. It either
skips them entirely (if non-root), or uses systemd-machined, or uses
cgroups directly.

This change adds ability to configure the mechanism for registering
resources between all these options explicitly. via

  <resource backend="none|cgroupfs|machined"/>

It is further possible to register directly with systemd and bypass
machined. We don't support this but systemd-nsspawn does and we ought
to consider this at some point. This would involve a new "systemd"
backend type alongside "machined".

Signed-off-by: Daniel P. Berrange <berrange@redhat.com>
Message-Id: <20200320134014.3123811-2-berrange@redhat.com>