qemuSecurityChownCallback: Don't initialize storage file subsystem for local file
virStorageFileSupportsSecurityDriver ends up initializing the storage
file backend which after the recent changes to the daemon architecture
may end up dlopening of the backend modules.

Since this is required only for remote storage we can optimize the call
by moving the check whether the backend is supported to the branch which
deals with remote storage. This will prevent the pointless dlopen for
local files.

This is amplified by the fact that we now do most storage driver
operations in forks so the storage backends may not even be initialized
in the main daemon. On the other hand, not having gluster linked in the
main daemon is a big win, so we should keep only the forks having to
link with it.

Signed-off-by: Peter Krempa <pkrempa@redhat.com>
Message-Id: <f4870568c39b9cebc4e6d2efda2d8bea5a97e13a.1585221445.git.pkrempa@redhat.com>
Reviewed-by: Pavel Mores <pmores@redhat.com>