gitlab: reduce number of cross build jobs run by default
Currently we nine different cross build jobs, but as we introduce more
native jobs this is going to result in a very long CI execution time.
For developers testing their personal branches under development it is
generally sufficient to just look at a couple of interesting scenarios,
namely 32-bit and big endian.

This splits the cross build jobs so that by default only the armv7
and s390x archs are built. The remainining archs are setup so that they
are only built for code on the master branch, which will have the effect
of doing post-merge testing. Developers can opt-in to full testing of
their pre-merge code by pushing it to a branch with a name prefix of
"ci-extra-".

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>
Message-Id: <20200326123538.4177062-6-berrange@redhat.com>