cpu_x86: Drop noTSX hint for incompatible CPUs
The hint was introduced a long time ago when broken TSX implementation
was found in Haswell and Broadwell CPUs. Since then many more CPUs with
TSX were introduced and and disabled due to TAA vulnerability.

Thus the hint is not very useful and I think removing it is a better
choice then updating it to cover all current noTSX models.

Signed-off-by: Jiri Denemark <jdenemar@redhat.com>
Message-Id: <da03eee1319592a2a254337deb50f3560a81e798.1585316580.git.jdenemar@redhat.com>