gitlab: Add and use 'environment' template
This template provides a standardized place where environment
variables, which might not be static but depend on the output of
some shell command, can be defined for later use. It's not very
interesting right now, but we'll put it to good use in a second.

This pattern is already used in libosinfo's GitLab CI integration.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>
Message-Id: <20200327171559.1109515-2-abologna@redhat.com>