Revert "remote: move timeout arg into sysconf file"
There is nothing really systemd-specific about passing extra
arguments to daemons so it's reasonable, although not currently the
case, that startup scripts written for other init systems might want
to source these sysconf files; for those init systems, which likely
do not support socket activation, making the daemon quit after a
timeout has expired is probably not a good idea.

More generally, the sysconf files should not reflect the default
behavior, but only contain overrides explicitly put in place by the
admin; now that we have a mechanism to disable timeouts regardless
of the default set in the service file, that argument for having the
default timeout in the sysconf file is moot as well.

This reverts commit 581767a98ab5f674ac335d6c270efa8576bfdfbf.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>
Message-Id: <20200401185345.248398-3-abologna@redhat.com>